% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{split_filename}
\alias{split_filename}
\title{Split filenames into path, name, and extension.}
\usage{
split_filename(filename)
}
\arguments{
\item{filename}{Character vector of filenames.}
}
\value{
A list with files for the \code{path}, \code{name}, and \code{ext} (extension).
}
\description{
Split filenames into path, name, and extension.
}
